gateway/requirements.txt
# =============================================================================
# Sentinel Gateway — Python dependencies
# Python 3.12 required
# =============================================================================

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Config
pydantic==2.7.1
pydantic-settings==2.2.1

# Redis (async)
redis[hiredis]==5.0.4

# Neo4j (async)
neo4j==5.20.0

# Policy YAML loading
PyYAML==6.0.1

# LLM clients (Phase 2 — listed here so gateway container has them)
anthropic==0.28.0
httpx==0.27.0

# Structured LLM output (Phase 2)
instructor==1.3.3

# LLM observability (Phase 2)
langfuse==2.36.1

# Circuit breaker (Phase 2)
pybreaker==1.2.0

# Retry logic
tenacity==8.3.0

# Logging
structlog==24.1.0

# Observability (Phase 4)
prometheus-client==0.20.0

# JWT (Phase 2)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
